+++ eslint.config.js
@@ -10,6 +10,9 @@ export default [
             globals: {
                 ...globals.browser,
             },
+            parserOptions: {
+                ecmaFeatures: { jsx: true },
+            },
         },
         rules: {
             'quotes': ['error', 'single', { avoidEscape: true }],